Fixing scripts wrt to incompatibility following full support for meta pattern unification in apply (i.e. even when args are variables of the context and not only locally bound variables).
Unfortunately, a bug in the recursive use of apply through conjunction
makes that the 2nd component of the open_neighborhood_basis_elements
field (of conclusion "In U a") is applied before the 1st component (of
conclusion "open U") which is the expected one (and which was actually
called before pattern unification allowed "In U a" to be used). Fixed
the problem by setting "In" opaque.

git-svn-id: svn://scm.gforge.inria.fr/svnroot/coq-contribs/trunk@1520 0cf17b13-060f-0410-b1b1-c666bec9822a