[B] Cardinals and well-orders are well-founded
The API for well-orders changes completely, an a lot of typeclasses are
used, so less boilerplate has to be written.

This statement enables/simplifies some arguments or constructions using
`Cardinals`. The same construction can be done in more general
environments like ETCS.

The proof is based on thoughts from
https://golem.ph.utexas.edu/category/2021/06/large_sets_3.html
and a lecture on set-theory in ZFC.

The proof (for cardinals) depends heavily on the well-ordering-theorem.
Are there models of ZF, where the order corresponding to `lt_cardinal`
and `le_cardinal` is not well-founded? I.e. there is some infinite
descending chain of strict injections.

Use a manual def. of `lexprod`, because it is not available in earlier
versions of coq (at least v8.12 to v8.14).