[B] Redefine `induced_function_correct`
Stating the lemma like this, the user doesn't have to use
proof-irrelevance to make his term equal to `quotient_projection R x`.
But on the other hand, the user of the lemma currently has to use
`unfold` on `quotient_projection R x` before he can do the rewrite.
This is less convenient and makes this change breaking.