Compact Hausdorff implies normal. Without choice
By juggling around, one can prove "compact Hausdorff spaces are normal"
without using choice. The relevant facts are, that "compactness is a
kind of choice property" and "finite sets have choice (exploitable by
induction)".

The lemma `compact_Hausdorff_impl_T3_sep_lemma` should get another name
and be moved to ZornsLemma.

The approach of the proof is pretty clear (whatever that means), and the
new proof has about the same length as the one based on choice (not
counting the new lemmas).

I created a new file `Finite_sets.v` (in allusion to the corresponding
file in the stdlib), because we don't have a file yet that is *only*
concerned with `Finite`. The best would have been `FiniteTypes.v`, but
this is all about `FiniteT`.