[B] Rewrite `completion_exists`
Split the proof into separate parts, one constructing the non-dense
completion, the other the dense completion.
Instead of forcing the resulting embedding to be dense w.r.t. to the
`MetricTopology` we allow the proof to provide an arbitrary topological
space equipped with a metric. This way, we don't need to manage the
metric topology or have to prove that all topological notions agree
between `MetricTopology` and a metrizable topology.

Note that the proof of density could be simplified a lot. The proof does
not need first-countability anymore for this part.