[B] Change def of complete metric
This is a breaking change, because some implicit arguments differ now,
and because `complete` does not require `metric _` as argument anymore.

The def. of complete metric now only depends on the metric and does not
mention any topology. Together with the characterisation w.r.t.
convergence, this suffices. The old definition was problematic, because
it is non-trivial to transfer convergence w.r.t. `MetricTopology` into
convergence w.r.t. an equivalent topology.