[B] Redefine eq_cardinal, invertible, FiniteT
Use `inverse_map` throughout, instead of `bijective`. This makes it
easier to re-use lemmas.

Use `invertible` instead of `bijective` because it is a slightly
stronger property. In particular `FiniteT` behaves a bit better with
`invertible` than with `bijective`.

Move the definition of `eq_cardinal` from Cardinals to
FiniteTypes, because to do it the the other way around, a some lemmas
need to be moved.