[B] Change order of Cardinals, Finite, Countable …
Moves the definition of `eq_cardinal` to Cardinals, where it belongs.
But to allow this, the statements `countable_img_inj` and
`FiniteT_cardinality` had to be moved to Countable and FiniteTypes
respectively. In addition, `CountableT_cardinality` was removed, because
`CountableT` could be redefined in terms of `le_cardinal`, making the
lemma even more trivial.

Also, removes the file InfiniteTypes, because it has become empty.

The lemmas `Finite_as_lt_cardinal_ens` and
`injective_finite_inverse_image` currently are in `FiniteTypes.v`,
because their proofs use `FiniteT` in essential ways.