[B] Edit imports and move finiteness lemmas
There were lemmas in `IndexedFamilies` and in `InverseImage` which
concerned `Finite` and `FiniteT`. Thus they had to depend on
`Finite_sets` and `FiniteTypes`.
This dependency makes it hard to develop a theory of cardinals which
uses `IndexedFamilies` and `InverseImage`, but then adds new facts about
`Finite` and `FiniteT` which should go into `Finite_sets` and
`FiniteTypes`.
So we remove the dependency by making the following changes:

* `finite_indexed_union` has been in `IndexedFamilies`. It is now in
  `FiniteTypes`.
* `Finite_sets` now contains lemmas concerning `inverse_image` instead
  of `InverseImage`
* `InverseImage` contains a new characterization of `injective` based on
  singletons.

This way, we reverse the direction of the dependency and make the way
free for a new formalisation of cardinalities.

In addition, we removed many unnecessary imports/exports, but had to add
new ones where they really were necessary. We also introduced a new
formatting for imports.