Introduce a notion `inverse_map`
Having this definition will simplify the redefinition of homeomorphism.
Using `inverse_map` in the def. of homeomorphism will make the proofs
more conceptual (fewer hypotheses to juggle around), but this will need
a lot of changes.

It would make sense to use `inverse_map` in the definition of
`invertible`, but there's a lot that would need to be changed. That's
why I postpone doing so.