Cleanup some imports
Some of the touched files depended unnecessarily on
`ZornsLemma.EnsemblesSpec`. Other dependencies could be removed as well.
On the other hand, some files now needed an `Import` of `EnsemblesSpec`,
because it isn't imported transitively anymore.

The imports were sometimes reformated to the new style.

In `Classical_Wf`, first the proof of `MEP_implies_WF` had to be
adapted. On the go, the proof was made more stable against name
mangling.

In `Topology.OpenBases` the proof of `coverable_by_open_basis_impl_open`
was adapted, so it didn't need `EnsemblesSpec` anymore.