Move Cardinals etc. into separate folder
The files concerning cardinalities (`eq_cardinal` and `le_cardinal`) in
general are rather large and would be easier to manage if they are split
into multiple files.
This also solves a "practical" problem: the statement of `CSB` could be
formulated using `eq_cardinal`. But this is not possible if the file
`Cardinals.v` depends on the file `CSB.v`.
Another possibility would be, that `CSB` is carefully stated, that it
uses definitions compatible with `Cardinals.v`. But this would break
`Search`, or we'd need to re-state `CSB` in terms of `eq_cardinal` and
`le_cardinal` in `Cardinals.v`.