Move `nat_unbounded_…` to CountableTypes.v
The lemma says "unbounded subsets of ℕ are countably infinite", which
does not belong to the general discussion of cardinals.
It is a bit hard to sort this lemma into the existing files, but
CountableTypes.v seems to work best, because it already is concerned
with `nat`.

Also moves `Finite_as_lt_cardinal_ens` and
`injective_finite_inverse_image` to CountableTypes.v, because they
depend on `nat_unbounded_impl_countably_infinite`. This is not ideal.

A bit problematic is `nat_minimal_element_property_dec` and
`nat_minimal_element_property` which are not exactly proofs of
`Classical_Wf.minimal_element_property`. And they are self-contained,
not depending on any theory in CountableTypes.v.