[B] Redo OrderTopology.v, define `orders_top`
Similar to `metrizes`, the predicate `orders_top` states, that a
topology on an arbitrary space coincides with an `OrderTopology`.
By stating the lemmas in `OrderTopology.v` in terms of `orders_top`
instead of `OrderTopology`, they become easier to apply to spaces which
are not defined as `OrderTopology`. For example `SubspaceTopology`.

Breaking changes: implicit arguments changed. Some proofs regarding
`RTop` had to be adapted.